refactor: move NodeStorage to storage
