+++ crates/storage/src/merkle_tree.rs
@@ -47,6 +47,84 @@ use rusqlite::{params, OptionalExtension, Transaction};
 
 use stark_hash::StarkHash;
 
+/// Backing storage for [`MerkleTree`].
+///
+/// Default implementation and persistent implementation is the `RcNodeStorage`. Testing/future
+/// implementations include [`HashMap`](std::collections::HashMap) and `()` based implementations
+/// where the backing storage is not persistent, or doesn't exist at all. The nodes will still be
+/// visitable in-memory.
+pub trait NodeStorage {
+    /// Find a persistent node during a traversal from the storage.
+    fn get(&self, key: StarkHash) -> anyhow::Result<Option<PersistedNode>>;
+
+    /// Insert or ignore if already exists `node` to storage under the given `key`.
+    ///
+    /// This does not imply incrementing the nodes ref count.
+    fn upsert(&self, key: StarkHash, node: PersistedNode) -> anyhow::Result<()>;
+
+    /// Decrement previously stored `key`'s reference count. This shouldn't fail for key not found.
+    #[cfg(test)]
+    fn decrement_ref_count(&self, key: StarkHash) -> anyhow::Result<()>;
+
+    /// Increment previously stored `key`'s reference count. This shouldn't fail for key not found.
+    fn increment_ref_count(&self, key: StarkHash) -> anyhow::Result<()>;
+}
+
+impl NodeStorage for () {
+    fn get(&self, _key: StarkHash) -> anyhow::Result<Option<PersistedNode>> {
+        // the rc<refcell> impl will do just fine by without any backing for transaction tree
+        // building
+        Ok(None)
+    }
+
+    fn upsert(&self, _key: StarkHash, _node: PersistedNode) -> anyhow::Result<()> {
+        Ok(())
+    }
+
+    #[cfg(test)]
+    fn decrement_ref_count(&self, _key: StarkHash) -> anyhow::Result<()> {
+        Ok(())
+    }
+
+    fn increment_ref_count(&self, _key: StarkHash) -> anyhow::Result<()> {
+        Ok(())
+    }
+}
+
+impl NodeStorage for std::cell::RefCell<std::collections::HashMap<StarkHash, PersistedNode>> {
+    fn get(&self, key: StarkHash) -> anyhow::Result<Option<PersistedNode>> {
+        Ok(self.borrow().get(&key).cloned())
+    }
+
+    fn upsert(&self, key: StarkHash, node: PersistedNode) -> anyhow::Result<()> {
+        use std::collections::hash_map::Entry::*;
+        if !matches!(node, PersistedNode::Leaf) {
+            match self.borrow_mut().entry(key) {
+                Vacant(ve) => {
+                    ve.insert(node);
+                }
+                Occupied(oe) => {
+                    let existing = oe.get();
+                    anyhow::ensure!(
+                        existing == &node,
+                        "trying to upsert a different node over existing? {existing:?} != {node:?}"
+                    );
+                }
+            }
+        }
+        Ok(())
+    }
+
+    #[cfg(test)]
+    fn decrement_ref_count(&self, _key: StarkHash) -> anyhow::Result<()> {
+        Ok(())
+    }
+
+    fn increment_ref_count(&self, _key: StarkHash) -> anyhow::Result<()> {
+        Ok(())
+    }
+}
+
 /// Provides a reference counted storage backend for the
 /// nodes of a Starknet Binary Merkle Patricia Tree.
 ///
@@ -158,7 +236,7 @@ impl Queries<'static> {
     }
 }
 
-impl<'a, 'queries> crate::state::merkle_tree::NodeStorage for RcNodeStorage<'a, 'queries> {
+impl<'a, 'queries> NodeStorage for RcNodeStorage<'a, 'queries> {
     fn get(&self, key: StarkHash) -> anyhow::Result<Option<PersistedNode>> {
         self.get(key)
     }
